Documentation badge.
Eventually, these badges will become old and 404 and it will look like
this project is entirely unmaintained and indeed, that will likely be
the case. There is a dizzying breadth of problems in open source
software regarding labour and this badge eventually 404ing and sticking
out something awful among the large row of badges is the true pinnacle
of demonstration. Yes, one day, this badge will 404. It's alt text will
ruin that shiny array. The documentation will stop building, people will
realize they don't have time to maintain these projects for free anymore.

But until then, I welcome this badge in the most triumphant manner, as a
symbol of my good faith that this project will not be like the others,
at least not in the short-term. (◡ ‿ ◡ ✿)